+++ navigator/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Add named navigation path
+* Add screen type navigation path
+* Add link based navigation support
+
 ## 0.1.0
 
 * Add routing configuration APIs